requirements.txt
fastapi
pydantic
streamlit
openai
tiktoken
numpy
pandas
python-dotenv
sqlite3